+++ .idea/misc.xml
@@ -4,9 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="../../../../layout/custom_preview.xml" value="0.1875" />
-        <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.138" />
         <entry key="app/src/main/res/drawable/ic_baseline_close.xml" value="0.2515" />
-        <entry key="app/src/main/res/drawable/ic_launcher_background.xml" value="0.138" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.25" />
         <entry key="app/src/main/res/layout/bottom_dialog.xml" value="0.3546875" />
         <entry key="app/src/main/res/layout/bottom_sheet.xml" value="0.2" />
