--- app/src/main/res/anim/rotate.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android">
-    <rotate android:fromDegrees="0"
-        android:toDegrees="360"
-        android:pivotX="50%"
-        android:pivotY="50%"
-        android:duration="600"
-        android:repeatMode="restart"
-        android:repeatCount="infinite"
-        android:interpolator="@android:anim/linear_interpolator"/>
-
-</set>
