+++ app/src/main/res/layout/activity_main.xml
@@ -1,126 +1,67 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".MainActivity">
+    tools:context=".MainActivity"
+    android:orientation="vertical">
 
-    <FrameLayout
-        android:id="@+id/fragment_container_view"
+    <androidx.camera.view.PreviewView
+        android:id="@+id/ScanPreview"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="0dp"
+        android:layout_weight="80" />
 
-        <Button
-            android:id="@+id/scan_button"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/Scan"
-            android:layout_gravity="center" />
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_weight="20"
+        android:orientation="vertical">
 
-        <LinearLayout
-            android:visibility="gone"
-            android:id="@+id/action"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center">
-            <View
-                android:layout_width="0dp"
-                android:layout_height="50dp"
-                android:layout_weight="1" />
-            <Button
-                android:id="@+id/outing_button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Outing">
-
-            </Button>
-            <Space
-                android:layout_width="0dp"
-                android:layout_height="1dp"
-                android:layout_weight="1" >
-            </Space>
-            <Button
-                android:id="@+id/food_button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-
-                android:text="Food">
-            </Button>
-            <View
-                android:layout_width="0dp"
-                android:layout_height="50dp"
-                android:layout_weight="1" />
-        </LinearLayout>
-
-
-        <LinearLayout
+        <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_weight="20"
-            android:orientation="vertical"
-            android:layout_gravity="bottom">
+            android:id="@+id/Name"
+            android:text="null"
+            android:textSize="15sp"
+            android:padding="2dp"
+            android:textAlignment="center"
 
-            <TextView
-                android:visibility="gone"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:id="@+id/Name"
-                android:text="Name"
-                android:textSize="15sp"
-                android:padding="2dp"
-                android:textAlignment="center"
-
-                />
-            <TextView
-                android:visibility="gone"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:id="@+id/RollNumber"
-                android:text="RollNumber"
-                android:textSize="15sp"
-                android:padding="2dp"
-                android:textAlignment="center"
+            />
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/RollNumber"
+            android:text="null"
+            android:textSize="15sp"
+            android:padding="2dp"
+            android:textAlignment="center"
 
 
-                />
-            <TextView
-                android:visibility="gone"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:id="@+id/Branch"
-                android:text="Branch"
-                android:textSize="15sp"
-                android:padding="2dp"
-                android:textAlignment="center"
+            />
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/Branch"
+            android:text="null"
+            android:textSize="15sp"
+            android:padding="2dp"
+            android:textAlignment="center"
 
 
-                />
-            <TextView
-                android:visibility="gone"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:id="@+id/Degree"
-                android:text="Degree"
-                android:textSize="15sp"
-                android:padding="2dp"
-                android:textAlignment="center"
+            />
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:id="@+id/Degree"
+            android:text="null"
+            android:textSize="15sp"
+            android:padding="2dp"
+            android:textAlignment="center"
 
-                />
-        </LinearLayout>
+            />
 
-
+    </LinearLayout>
 
 
-    </FrameLayout>
-
-
-
-
-
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
-
-
-
+    </LinearLayout>
