+++ app/src/main/res/layout/fragment_bottom_sheet.xml
@@ -1,22 +1,25 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
+    android:background="#546471"
     tools:context=".BottomSheet">
 
-    <androidx.camera.view.PreviewView
-        android:id="@+id/ScanPreview"
-        android:layout_width="match_parent"
-        android:layout_height="600dp">
+    <Button
+        android:id="@+id/button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="100dp"
+        android:layout_marginStart="100dp"
 
-    </androidx.camera.view.PreviewView>
+        android:text="@string/outing" />
 
-
-
-
-
-
-
-</LinearLayout>
+    <Button
+        android:id="@+id/button2"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="100dp"
+        android:layout_marginStart="250dp"
+        android:text="@string/food" />
+</FrameLayout>
