+++ app/src/main/res/values/strings.xml
@@ -7,5 +7,4 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="food">Food</string>
     <string name="outing">Outing</string>
-    <string name="Scan">Scan ID Card</string>
 </resources>
