+++ build.gradle
@@ -1,6 +1,6 @@
 buildscript {
     dependencies {
-        classpath 'com.google.gms:google-services:4.3.14'
+        classpath 'com.google.gms:google-services:4.3.10'
     }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
